Move fake-data creators to  separate`test_utils` folder.
