+++ nupic/research/frameworks/pytorch/dataset_utils/__init__.py
@@ -19,5 +19,4 @@
 #
 
 from .dataset_utils import *
-from .fake_data import *
 from .hdf5_utils import *
