+++ nupic/research/frameworks/pytorch/test_utils/fake_data.py
