+++ tests/unit/frameworks/pytorch/imagenet_experiment_test.py
@@ -23,11 +23,9 @@ import unittest
 
 import torch
 
-from nupic.research.frameworks.pytorch.imagenet import (
-    ImagenetExperiment,
-    TempFakeSavedData,
-)
+from nupic.research.frameworks.pytorch.imagenet import ImagenetExperiment
 from nupic.research.frameworks.pytorch.models.resnets import resnet50
+from nupic.research.frameworks.pytorch.test_utils import TempFakeSavedData
 
 
 class ImagenetExperimentTest(unittest.TestCase):
